+++ runtime/lib/descriptor/activity-descriptor.js
@@ -656,52 +656,6 @@ Object.assign(ActivityDescriptor.prototype,
       fn: function getMicMute () {
         return Promise.resolve(this._runtime.component.turen.muted)
       }
-    },
-    /**
-     * begin tts mix with audio.
-     *
-     * @memberof yodaRT.activity.Activity
-     * @instance
-     * @function mixTtsBegin
-     * @param {interrupt} [interrupt] interrupt audio if true, otherwise suppress audio. Undefined will use system config.
-     * @param {playerId} [number] playerId is required if the value of interrupt is false.
-     * @returns {Promise} Promise of audio mixed
-     */
-    mixTtsBegin: {
-      type: 'method',
-      returns: 'promise',
-      fn: function mixTtsBegin (interrupt, playerId) {
-        // case: mixTtsBegin()
-        if (interrupt === undefined) {
-          playerId = -1
-        }
-        // case: mixTtsBegin(playerId)
-        if (typeof interrupt === 'number') {
-          playerId = interrupt
-          interrupt = undefined
-        }
-        // case: mixTtsBegin(interrupt)
-        if (playerId === undefined) {
-          playerId = -1
-        }
-        return this._runtime.component.audioMix.mixTtsBegin(this._appId, playerId)
-      }
-    },
-    /**
-     * end audio mix.
-     *
-     * @memberof yodaRT.activity.Activity
-     * @instance
-     * @function mixTtsEnd
-     * @returns {Promise} Promise of audio mix end
-     */
-    mixTtsEnd: {
-      type: 'method',
-      returns: 'promise',
-      fn: function mixTtsEnd () {
-        this._runtime.component.audioMix.mixTtsEnd()
-        return Promise.resolve()
-      }
     }
   }
 )
